Backport adaptations
